+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Added new command `entrypoint` that is available only in version for docker images,
+  that is used as *entrypoint* for docker container and that is responsible for applying
+  configuration from environment variables to Odoo configuration file before
+  any further action.
+
+
+---
+
 ## Release 0.2.0 (2024-12-12)
 
 ### Added